refactor formatter
refactored the formatter to avoid having to write a bunch of setup stuff a bunch of times.

file loading and parsing is now handled in the main  file (index.ts). previously individual formatting functions that used to load and parse files, are now “output structure template” functions.

also updated the input data to include a couple other keys.

Signed-off-by: Samantha Emily-Rachel Belnavis <samanthabelnavis@gmail.com>